refactor: refactor structure of page-not-found
