+++ src/pages/page-not-found/index.js
@@ -1,0 +1,1 @@
+export { default as PageNotFound } from './page-not-found';
